robotik_chess_stack/agent/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pyyaml==6.0.2
requests==2.32.3
numpy==1.26.4
